CORE-7943 Centralize consensual ledger integration test setup code
