+++ components/ledger/ledger-consensual-flow/build.gradle
@@ -42,9 +42,6 @@ dependencies {
     testImplementation project(':libs:serialization:serialization-kryo')
     testImplementation project(':testing:test-serialization')
 
-    integrationTestImplementation project(':testing:sandboxes')
-    integrationTestImplementation project(':testing:sandboxes-testkit')
-
     integrationTestRuntimeOnly "org.apache.aries.spifly:org.apache.aries.spifly.dynamic.framework.extension:$ariesDynamicFrameworkExtensionVersion"
     integrationTestRuntimeOnly project(':libs:messaging:db-message-bus-impl')
     integrationTestRuntimeOnly project(':libs:messaging:messaging-impl')
